# docs/api/05-openapi.yaml
openapi: 3.1.0
info:
  title: EventFlow API
  version: 1.0.0
  description: |
    Contrato oficial Backend ↔ Android de EventFlow (API First, fuente única de verdad).
    Convenciones: docs/api/01-api-guidelines.md · Errores RFC 9457: docs/api/02-api-error-model.md ·
    Paginación keyset: docs/api/07-pagination-filtering.md. Dinero como string decimal, fechas ISO-8601 UTC,
    enums UPPER_SNAKE (los clientes DEBEN tolerar valores desconocidos).
  contact:
    name: EventFlow Architecture
  license:
    name: Proprietary
    identifier: LicenseRef-EventFlow-Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo
  - url: https://api.eventflow.app/api/v1
    description: Producción (placeholder)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: Autenticado sin permiso (incluye qr_not_yet_visible)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicto de estado o concurrencia
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: Validación o regla de negocio
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit excedido (ver Retry-After)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    FieldError:
      type: object
      required: [field, message, code]
      properties:
        field: { type: string }
        message: { type: string }
        code: { type: string }
    Problem:
      type: object
      description: RFC 9457 + extensiones EventFlow (docs/api/02-api-error-model.md)
      required: [type, title, status, detail, instance, code, timestamp, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: ErrorCode máquina — clave de enrutamiento del cliente }
        timestamp: { type: string, format: date-time }
        traceId: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        retryAfterSeconds: { type: integer }
        conflictVersion: { type: integer }

    # ===== Enums (los clientes toleran valores desconocidos) =====
    RoleCode: { type: string, enum: [ADMIN, ORGANIZER, STAFF, ATTENDEE] }
...
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 8, maxLength: 72 }
        fullName: { type: string, minLength: 1, maxLength: 200 }
        phone: { type: string, pattern: '^\+[1-9]\d{6,14}$' }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    UserProfile:
      type: object
      required: [id, email, fullName, roles, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        phone: { type: [string, 'null'] }
        roles:
          type: array
          items: { $ref: '#/components/schemas/RoleCode' }
        createdAt: { type: string, format: date-time }
    AuthTokensResponse:
      type: object
      required: [accessToken, accessTokenExpiresIn, refreshToken, user]
      properties:
        accessToken: { type: string }
        accessTokenExpiresIn: { type: integer, description: Segundos de vida del access token }
        refreshToken: { type: string }
        user: { $ref: '#/components/schemas/UserProfile' }
    UpdateProfileRequest:
      type: object
      required: [fullName]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        phone: { type: [string, 'null'], pattern: '^\+[1-9]\d{6,14}$' }
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Registro de asistente
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AuthTokensResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Inicio de sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example: { email: ana@mail.com, password: 'S3gura!pass' }
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AuthTokensResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Rotación de refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Tokens rotados (el refresh anterior queda inválido)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AuthTokensResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoca el refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '204': { description: Sesión cerrada }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ======================= CATALOG =======================
...
  /me:
    get:
      tags: [me]
      operationId: getProfile
      summary: Mi perfil
      responses:
        '200':
          description: Perfil
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/UserProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [me]
      operationId: updateProfile
      summary: Actualiza mi perfil
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/UserProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...